são 19:25... comecei às 15:12.... ao menos as leaks e os invalid frees desapareceram.

+++ dom/tree.c
@@ -195,6 +195,8 @@ static void __destroy_tree(tree_node* root){
 }
 
 void destroy_tree(tree_root* root){
+  if(root == NULL || root->root == NULL)
+    return;
   __destroy_tree(root->root);
   free(root);
 }
